+++ lib/need_for_type/states/end.rb
@@ -1,12 +1,9 @@
-require 'curses'
-require 'json'
-require 'httparty'
-
 require 'need_for_type/states'
+require 'need_for_type/api'
 
 module NeedForType::States
   class End < Menu
-    @@api_url = 'localhost:3000'
+    include NeedForType::API
 
     def initialize(display_window, stats, difficulty, text_id)
       super(display_window)
@@ -15,30 +12,24 @@ module NeedForType::States
       @text_id = text_id
 
       @option = 0
-      @scores = get_scores
+      @scores = get_scores(@text_id)
     end
 
     def update
       @display_window.render_end(@stats, @scores, @option)
 
-      input_worker(3) do
+      input_worker(4) do
         case @option
         when 0
-          return NeedForType::States::Game.new(@display_window, @difficulty)
+          return NeedForType::States::SubmitScore.new(@display_window, self, @text_id, @stats)
         when 1
+          return NeedForType::States::Game.new(@display_window, @difficulty)
+        when 2
           return NeedForType::States::Start.new(@display_window)
         end
       end
 
       return self
-    end
-
-    private
-
-    def get_scores
-      url = "http://#{@@api_url}/v1/scores?text_id=#{@text_id}"
-      response = HTTParty.get(url)
-      JSON.parse(response.body)
     end
   end
 end