+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Top scores (API) rendering on the end state 
+- SubmitScore state to submit score to server
 
 ## Version 0.2.2 - 2016/09/29